+++ src/main/kotlin/person/Person.kt
@@ -1,0 +1,7 @@
+package person
+
+data class Person(
+    val name: String,
+    val age: Int,
+    var nickname: String? = "Guest"
+)
