+++ src/main/kotlin/racingame/GameApplication.kt
@@ -1,0 +1,10 @@
+package racingame
+
+import racingame.controller.GameController
+import racingame.view.InputView
+import racingame.view.OutputView
+
+fun main() {
+    val gameController = GameController(InputView(), OutputView())
+    gameController.start()
+}
