+++ src/main/kotlin/racingame/domain/CarMoveCondition.kt
@@ -1,0 +1,15 @@
+package racingame.domain
+
+class CarMoveCondition(
+    private val numberGenerator: NumberGenerator = RandomNumberGenerator()
+) : MoveCondition {
+
+    override fun isMovable(): Boolean {
+        val number = numberGenerator.generate()
+        return number >= FORWARD_CONDITION
+    }
+
+    companion object {
+        const val FORWARD_CONDITION = 4
+    }
+}
