+++ src/main/kotlin/racingame/domain/Cars.kt
@@ -1,0 +1,12 @@
+package racingame.domain
+
+class Cars(
+    values: List<Car> = emptyList()
+) {
+    var values = values
+        private set
+
+    fun add(car: Car) {
+        values = values + Car(name = car.name, position = car.position)
+    }
+}
