+++ src/main/kotlin/racingame/domain/MoveCondition.kt
@@ -1,0 +1,5 @@
+package racingame.domain
+
+interface MoveCondition {
+    fun isMovable(): Boolean
+}
