+++ src/main/kotlin/racingame/domain/NumberGenerator.kt
@@ -1,0 +1,5 @@
+package racingame.domain
+
+interface NumberGenerator {
+    fun generate(): Int
+}
