--- src/main/kotlin/step1/Person.kt
@@ -1,7 +1,0 @@
-package step1
-
-data class Person(
-    val name: String,
-    val age: Int,
-    var nickname: String? = "Guest"
-)
