--- src/main/kotlin/step3/domain/CarMoveCondition.kt
@@ -1,15 +1,0 @@
-package step3.domain
-
-class CarMoveCondition(
-    private val numberGenerator: NumberGenerator = RandomNumberGenerator()
-) : MoveCondition {
-
-    override fun isMovable(): Boolean {
-        val number = numberGenerator.generate()
-        return number >= FORWARD_CONDITION
-    }
-
-    companion object {
-        const val FORWARD_CONDITION = 4
-    }
-}
