--- src/main/kotlin/step3/domain/MoveCondition.kt
@@ -1,5 +1,0 @@
-package step3.domain
-
-interface MoveCondition {
-    fun isMovable(): Boolean
-}
