--- src/main/kotlin/step3/domain/NumberGenerator.kt
@@ -1,5 +1,0 @@
-package step3.domain
-
-interface NumberGenerator {
-    fun generate(): Int
-}
