--- src/main/kotlin/step3/domain/RandomNumberGenerator.kt
@@ -1,16 +1,0 @@
-package step3.domain
-
-import java.util.Random
-
-class RandomNumberGenerator : NumberGenerator {
-    private val randomNumber: Int
-        get() = Random().nextInt(MAX)
-
-    override fun generate(): Int {
-        return randomNumber
-    }
-
-    companion object {
-        private const val MAX = 10
-    }
-}
