+++ src/test/kotlin/person/StringTest.kt
@@ -1,0 +1,12 @@
+package person
+
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Test
+
+class StringTest {
+    @Test
+    fun isEmpty() {
+        assertThat("".isEmpty()).isTrue()
+        assertThat(" ".isEmpty()).isFalse()
+    }
+}
