+++ src/test/kotlin/racingame/fake/FakeNumberGenerator.kt
@@ -1,0 +1,9 @@
+package racingame.fake
+
+import racingame.domain.NumberGenerator
+
+class FakeNumberGenerator(private val number: Int) : NumberGenerator {
+    override fun generate(): Int {
+        return number
+    }
+}
