--- src/test/kotlin/step1/StringTest.kt
@@ -1,12 +1,0 @@
-package step1
-
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
-
-class StringTest {
-    @Test
-    fun isEmpty() {
-        assertThat("".isEmpty()).isTrue()
-        assertThat(" ".isEmpty()).isFalse()
-    }
-}
