--- src/test/kotlin/step3/fake/FakeNumberGenerator.kt
@@ -1,9 +1,0 @@
-package step3.fake
-
-import step3.domain.NumberGenerator
-
-class FakeNumberGenerator(private val number: Int) : NumberGenerator {
-    override fun generate(): Int {
-        return number
-    }
-}
